# omod/src/main/webapp/resources/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "REST API for Visits Module.",
    "version" : "v1",
    "title" : "Visits Module REST API"
  },
  "basePath" : "/openmrs/ws",
...
  "paths" : {
...
    "/visits/overview/{uuid}" : {
      "get" : {
        "tags" : [ "REST API for visit overview" ],
        "summary" : "Get visits for location",
        "description" : "Get visits for location",
        "operationId" : "getVisitsForLocation",
        "parameters" : [ {
          "name" : "uuid",
          "in" : "path",
          "description" : "uuid",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "query",
          "in" : "query",
          "description" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "visitStatus",
          "in" : "query",
          "description" : "visitStatus",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "dateFrom",
          "in" : "query",
          "description" : "dateFrom",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "dateTo",
          "in" : "query",
          "description" : "dateTo",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "timePeriod",
          "in" : "query",
          "description" : "timePeriod",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "visits for location retrieved",
            "schema" : {
              "$ref" : "#/definitions/OverviewDTO"
            }
          },
          "404" : {
            "description" : "Location not found"
          },
          "500" : {
            "description" : "Visits for location not retrieved"
          }
        }
      }
    },
...
  },
  "definitions" : {
    "NameUrlDTO" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "url" : {
          "type" : "string"
        }
      }
    },
    "OverviewDTO" : {
      "type" : "object",
      "properties" : {
        "uuid" : {
          "type" : "string"
        },
        "patientIdentifier" : {
          "type" : "string"
        },
        "nameUrl" : {
          "$ref" : "#/definitions/NameUrlDTO"
        },
        "startDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "time" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        },
        "status" : {
          "type" : "string"
        },
        "location" : {
          "type" : "string"
        },
        "actualDate" : {
          "type" : "string",
          "format" : "date-time"
        }
      }
    },
...
  }
}